+++ src/components/Resume/Template3/index.tsx
@@ -14,7 +14,7 @@ import {
 } from '@ant-design/icons';
 import _ from 'lodash-es';
 import { FormattedMessage, useIntl } from 'react-intl';
-import { getDefaultTitleNameMap } from '@/datas/constant';
+import { getDefaultTitleNameMap } from '@/data/constant';
 import type { ResumeConfig, ThemeConfig } from '../../types';
 import './index.less';
