+++ src/helpers/store-to-local.ts
@@ -3,7 +3,7 @@ import { message } from 'antd';
 import type { ResumeConfig } from '@/components/types';
 import { customAssign } from '@/helpers/customAssign';
 import _ from 'lodash-es';
-import { RESUME_INFO } from '@/datas/resume';
+import { RESUME_INFO } from '@/data/resume';
 import { fetchResume } from './fetch-resume';
 import { intl } from '@/i18n';
